refactor: move customElements implementation to a separate module
