+++ scripts/build.js
@@ -3,7 +3,7 @@ const {mkdir, writeFile} = require('fs');
 const {dirname, resolve} = require('path');
 const {rollup} = require('rollup');
 const {promisify} = require('util');
-const config = require('../rollup.config');
+const configs = require('../rollup.config');
 
 const mkdirAsync = promisify(mkdir);
 const writeFileAsync = promisify(writeFile);
@@ -13,19 +13,25 @@ const detection = '  if (detect()) { return; }';
 
 (async () => {
   try {
-    const bundle = await rollup(config);
-    const generated = await bundle.generate(config.output);
-    const {
-      output: [{code}],
-    } = generated;
+    await Promise.all(
+      configs.map(async (config, index) => {
+        const bundle = await rollup(config);
+        const generated = await bundle.generate(config.output);
+        let {
+          output: [{code}],
+        } = generated;
 
-    const [first, second, ...other] = code.split('\n');
-    const result = [first, second, detection, ...other].join('\n');
+        if (index === 0) {
+          const [first, second, ...other] = code.split('\n');
+          code = [first, second, detection, ...other].join('\n');
+        }
 
-    const resultFile = resolve(cwd, config.output.file);
+        const resultFile = resolve(cwd, config.output.file);
 
-    await mkdirAsync(dirname(resultFile), {recursive: true});
-    await writeFileAsync(resultFile, result, 'utf8');
+        await mkdirAsync(dirname(resultFile), {recursive: true});
+        await writeFileAsync(resultFile, code, 'utf8');
+      }),
+    );
   } catch (e) {
     console.error(e.stack);
     process.exit(1);
