+++ src/shared.js
@@ -137,16 +137,3 @@ export const nativeConstructorNames = [
   'UList',
   'Video',
 ];
-
-if (!('customElements' in window)) {
-  const mock = () => {
-    throw new Error('Not supported');
-  };
-
-  window.customElements = {
-    define: mock,
-    get: mock,
-    upgrade: mock,
-    whenDefined: mock,
-  };
-}
